requirements.txt
altair==5.2.0
amqp==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backoff==2.2.1
bcrypt==4.1.1
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.4
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.0
constantly==23.10.4
contourpy==1.2.0
cron-descriptor==1.4.0
cryptography==41.0.7
cycler==0.12.1
dacite==1.8.1
daphne==4.0.0
dateutils==0.6.12
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-auditlog==2.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-import-export==3.3.1
django-redis==5.4.0
django-session-timeout==0.1.0
django-simpleui==2023.11.16
django-timezone-field==6.1.0
dnspython==2.4.2
duckdb==0.9.2
et-xmlfile==1.1.0
eventlet==0.33.3
executing==2.0.1
flower==2.0.1
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
gw_dsl_parser==0.1.38
htmlmin==0.1.12
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
ImageHash==4.3.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
incremental==22.10.0
ipython==8.18.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyterlab-widgets==3.0.9
kanaries_track==0.0.3
kiwisolver==1.4.5
kombu==5.3.4
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
multimethod==1.10
mysqlclient==2.2.0
networkx==3.2.1
numba==0.58.1
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
paramiko==3.3.1
parso==0.8.3
patsy==0.5.4
pexpect==4.9.0
phik==0.12.3
Pillow==10.1.0
plotly==5.18.0
ply==3.11
prettytable==3.9.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pyecharts==2.0.4
Pygments==2.17.2
pygwalker==0.3.18
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypng==0.20220715.0
pysmi==0.3.4
pysnmp==4.4.12
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
referencing==0.32.0
reportlab==4.0.8
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scipy==1.11.4
seaborn==0.12.2
segment-analytics-python==2.2.3
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
sqlglot==20.3.0
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.29.0
streamlit-aggrid==0.3.4.post3
streamlit-ydata-profiling==0.2.1
sweetviz==2.3.1
tablib==3.5.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
Twisted==23.10.0
txaio==23.1.1
typeguard==4.1.5
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
vine==5.1.0
visions==0.7.5
wasmtime==12.0.0
watchdog==3.0.0
wcwidth==0.2.12
widgetsnbextension==4.0.9
wordcloud==1.9.3
xlrd==2.0.1
xlwt==1.3.0
ydata-profiling==4.6.3
zipp==3.17.0
zope.interface==6.1
